lesslol/JadeCore548-patched

+++ src/server/scripts/Pandaria/ThroneOfThunder/boss_council.cpp
@@ -421,7 +421,7 @@ public:
         EventMap events;
 
         bool intro;
-        float damage;
+        uint32 damage;
         int32 hppcts;
         Unit* target;
 
@@ -829,7 +829,7 @@ public:
 
         bool intro;
         int32 hppcts;
-        float damage;
+        uint32 damage;
 
         void Reset()
         {
@@ -1201,7 +1201,7 @@ public:
         bool intro;
         int32 hppcts;
         bool explode;
-        float damage;
+		uint32 damage;
 
         void Reset()
         {
@@ -1565,8 +1565,8 @@ public:
 
         bool intro;
         bool explode;
-        float damage;
-        int32 timepertick;
+        uint32 damage;
+        uint32 timepertick;
         int32 hppcts;
         Unit* target = NULL;
 
@@ -2000,7 +2000,7 @@ public:
 
         InstanceScript* instance;
         EventMap events;
-        int32 jumpTime;
+        uint32 jumpTime;
         Creature* target = NULL;
         std::list<Creature*> councilBosses;
         bool contact;
@@ -2095,7 +2095,7 @@ public:
 
         InstanceScript* pInstance;
         EventMap events;
-        int32 jumpTime;
+        uint32 jumpTime;
         Player* target = NULL;
 
         void Reset()
@@ -2168,8 +2168,8 @@ public:
 
         InstanceScript* pInstance;
         EventMap events;
-        int32 ticktime;
-        int32 slowticktime;
+        uint32 ticktime;
+        uint32 slowticktime;
 
         void Reset()
         {
